requirements.txt
godirect
pylsl
vernierpygatt